Update db_model.py
I also added the session id to be uuid  also switched it to get the connection string for postgresql from my .env file to use postgre

+++ db_model.py
@@ -1,20 +1,25 @@
-# models.py
 from sqlalchemy import create_engine, Column, Integer, Boolean, String, ForeignKey, DateTime, func
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, scoped_session
+from dotenv import load_dotenv
+import os
+from sqlalchemy.dialects.postgresql import UUID
+import uuid
+
+load_dotenv()
 
 Base = declarative_base()
 
 class ValidBadgeIDs(Base):
     __tablename__ = 'valid_badge_ids'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    session_id = Column(String, ForeignKey('sessions.id'))
+    session_id = Column(UUID(as_uuid=True), ForeignKey('sessions.session_id'), nullable=False)
     badge_id = Column(String)
 
 class Ballot(Base):
     __tablename__ = 'ballots'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    session_id = Column(String, ForeignKey('sessions.id'))
+    session_id = Column(UUID(as_uuid=True), ForeignKey('sessions.session_id'), nullable=False)
     badge_id = Column(String)
     badge_status = Column(String)
     s3_key = Column(String)
@@ -25,27 +30,27 @@ class Ballot(Base):
 class UploadedZip(Base):
     __tablename__ = 'uploaded_zips'
     id = Column(Integer, primary_key=True, autoincrement=True)
-    session_id = Column(String)
+    session_id = Column(UUID(as_uuid=True), ForeignKey('sessions.session_id'), nullable=False)
     filename = Column(String)
 
 class UserSession(Base):
     __tablename__ = 'sessions'
-    id = Column(Integer, primary_key=True)
-    session_id = Column(String)
-    password = Column(String)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    session_id = Column(UUID(as_uuid=True), unique=True, nullable=False, default=uuid.uuid4)
+    password = Column(String, nullable=False)
 
 class OCRResult(Base):
     __tablename__ = 'ocr_results'
-    id = Column(Integer, primary_key=True, autoincrement = True)
-    session_id = Column(String, ForeignKey('sessions.id'))
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    session_id = Column(UUID(as_uuid=True), ForeignKey('sessions.session_id'), nullable=False)
     filename = Column(String)
-    extracted_text = Column(String)  # for debugging
+    extracted_text = Column(String)
 
 class BallotVotes(Base):
     __tablename__ = 'votes'
     id = Column(Integer, primary_key=True, autoincrement=True)
     badge_id = Column(String)
-    ballot_id = Column(Integer, ForeignKey('ballots.id'))
+    ballot_id = Column(Integer, ForeignKey('ballots.id'), nullable=False)
     name = Column(String)
     category_id = Column(String)
     vote = Column(String)
@@ -53,7 +58,14 @@ class BallotVotes(Base):
     vote_status = Column(String)
     is_valid = Column(Boolean)
 
-engine = create_engine('sqlite:///data.db')
+
+DATABASE_URL = os.getenv("DATABASE_URL")
+
+if not DATABASE_URL:
+    raise ValueError("DATABASE_URL not defined in .env file")
+
+engine = create_engine(DATABASE_URL)
+
 Base.metadata.drop_all(bind=engine)
 Base.metadata.create_all(engine)
